+++ hub-client/src/locales/nl.ts
@@ -1,6 +1,5 @@
 const nl = {
 	state: {
-		no_access: 'Je browser geeft geen toegang tot de cookies en lokale opslag van de Hub client. Waarschijnlijk omdat de browser de Hub client aanziet voor een `third-party tracker`. Druk op de knop om toegang te verlenen.',
 		button_request: 'Verleen toegang',
 		initial: 'Controleren of de hub client toegang heeft tot cookies en lokale opslag...',
 		requesting: 'Toegang verlenen...',